requirements.txt
requests
python-telegram-bot
rss_parser==1.2.1
python-telegram-bot[job-queue]
beautifulsoup4